ride model update

+++ server/Models/Ride.cs
@@ -25,6 +25,6 @@ namespace Server.Models
         public int FreeTickets { set; get; }
         
         [Required]
-        public int Price { set; get; }
+        public double Price { set; get; }
     }
 }
